pyproject.toml
[project]
name = "elearning-graphene-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "asgiref==3.8.1",
    "django==5.2",
    "django-redis>=6.0.0",
    "graphene==3.4.3",
    "graphene-django==3.2.3",
    "graphql-core==3.2.6",
    "graphql-relay==3.2.0",
    "pillow>=11.3.0",
    "promise==2.3",
    "psycopg==3.2.6",
    "psycopg-binary==3.2.6",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pyjwt==2.10.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "six==1.17.0",
    "sqlparse==0.5.3",
    "text-unidecode==1.3",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "uvicorn>=0.35.0",
]
